remove load methods abstraction
