--- packages/insomnia/src/network/grpc/proto-loader/__mocks__/index.ts
@@ -1,7 +1,0 @@
-import { jest } from '@jest/globals';
-
-module.exports = {
-  loadMethods: jest.fn(),
-  loadMethodsFromPath: jest.fn(),
-  getSelectedMethod: jest.fn(),
-};
